+++ git_p4son/__init__.py
@@ -7,6 +7,6 @@ creating and updating changelists, and managing Swarm reviews.
 
 CONFIG_DIR = '.git-p4son'
 
-__version__ = "0.2.11"
+__version__ = "0.3.0"
 __author__ = "Andreas Andersson"
 __email__ = "andreas@neoboid.com"